fix: migration

+++ db/migrate/20230604153339_create_blog_comments.rb
@@ -1,7 +1,7 @@
 class CreateBlogComments < ActiveRecord::Migration[6.1]
   def change
     create_table :blog_comments do |t|
-      t.references :end_user, type: :integer, null: false, foreign_key: true
+      t.references :end_user, type: :bigint, null: false, foreign_key: true
       t.references :post_blog, type: :integer, null: false, foreign_key: true
       t.text :comment,    null: false
